+++ src/model/investment.model.js
@@ -1,7 +1,8 @@
 import knex from "../db";
-import { jsonObjectQuerySelect } from "../utils";
+import { jsonObjectArrayQuerySelect, jsonObjectQuerySelect } from "../utils";
 import transacting from "../utils/transacting";
 import * as categoryModel from "./category.model";
+import * as transactionModel from "./transaction.model";
 
 const TABLE_NAME = "investment";
 export const selectDefault = [
@@ -40,6 +41,28 @@ export const selectDefault = [
  * @property {String} updatedAt
  */
 
+/**
+ * @param {Object} options 
+ * @param {Investment} options.where 
+ * @param {number} options.limit 
+ * @param {import('knex').Knex.Transaction} trx 
+ * @returns {import('knex').Knex.QueryBuilder}
+ */
+export const findTransaction = (options, trx) => {
+    const query = knex(TABLE_NAME)
+        .select([
+            knex.raw(jsonObjectArrayQuerySelect("transaction", transactionModel.selectDefault)),
+            ...selectDefault.map((select) => {
+                return `${TABLE_NAME}.${select}`;
+            })
+        ])
+        .innerJoin("transaction", "transaction.investmentId", "=", `${TABLE_NAME}.id`)
+        .whereNotNull("transaction.id")
+        .where(options.where)
+        .limit(options.limit || 10)
+        .groupBy(`${TABLE_NAME}.id`);
+    return transacting(query, trx);
+};
 /**
  * @param {Object} options 
  * @param {Investment} options.where